requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-socketio>=5.11
starlette>=0.36
requests>=2.31
httpx>=0.27
redis>=5.0
